Hide jstris stats until 4 lines remaining
Bookmarklet to hide `lrem` and `gstats` div IDs at the start and unhide just
`lrem` (number of lines remaining) when there are 4 or fewer left.

At the moment, this won't show the final time in `gstats` when the game ends,
so the only way to see the result is to look in the chat. It also won't yet
re-hide the elements if you restart the game or play another game.

`lrem` also gets un-hidden without any feedback, so it may be hard to notice.

Signed-off-by: Tushar Pankaj <tushar.s.pankaj@gmail.com>